requirements.txt
fastapi
SQLAlchemy
transformers
sentence_transformers
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0
nltk
sentencepiece
protobuf
websockets
psycopg2
asyncpg
python-dotenv
PyJWT
redis
celery
numpy==1.24.1